% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTop10.R
\name{topten}
\alias{topten}
\title{Building a Model with Top Ten Features}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a n x p matrix of n observations and p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a 'lm' object representing the linear model with the top 10 predictors
}
\description{
This function develops a prediction algorithm based on the top 10 features
in 'x' that are most predictive of 'y'.
}
\details{
This function runs a univariate regression of y on each predictor in x and
calculates the p-value indicating the significance of the association. The
final set of 10 predictors is the taken from the 10 smallest p-values.
}
\author{
Todor
}
\seealso{
\code{lm}
}

